Android: Refactor size maintenance to own class.
